Handle async BT stack better
Previous version of the code assumed command to
set HCI_POWER_OFF happened right away. It is a
process that terminates with an event. Handling
this as an event makes switch between client and
server modes work better.

There are still issues switching back and forth,
but it is better. More testing is needed to figure
it out.